2nd Prob/requirements.txt
# Contract Analysis & Risk Assessment Bot - Dependencies
# GUVI HCL Hackathon 2026

# Core Framework
streamlit>=1.30.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0

# NLP
spacy>=3.7.0
nltk>=3.8.0

# LLM Integration (GPT-4)
openai>=1.0.0

# Hindi Language Support
googletrans==4.0.0-rc1

# PDF Export
fpdf2>=2.7.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0
